Rearranged some tests to make more sense
