+++ test/dom/hook-events.js
