+++ test/dom/mount-hook.js
