+++ test/dom/render-hook.js
