+++ test/dom/update-attributes.js
