+++ test/dom/update-elements.js
