+++ test/dom/update-hook.js
