+++ test/dom/update-replace.js
