+++ test/dom/update-text.js
