+++ test/string/index.js
